Refactors admin functionalities into separate controllers
Removes AdminController and delegates its responsibilities to dedicated controllers: UserController, ConferenceController, EditorController, and PageController. Updates API routes to reflect the new structure, improving modularity and maintainability.

No functional changes were introduced.
